DDIM-jittor/requirements.txt
astunparse==1.6.3
certifi==2025.6.15
charset-normalizer==3.4.2
idna==3.10
jittor==1.3.9.14
lmdb==1.6.2
numpy==1.24.0
nvidia-ml-py==12.575.51
pandas==2.3.0
pillow==11.3.0
pynvml==12.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
six==1.17.0
tqdm==4.67.1
tzdata==2025.2
urllib3==2.5.0
